move pool to its own file
